// src/app/models/character.model.ts
import { Origin } from './origin.model';

export interface Character {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: Origin;
  location: Location;
  image: string;
  episode: any[];
  url: string;
  created: Date;
  pilotEpisode: string;
}